Rename file with hyphen as per coding standards
